# Dealer Adapt API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Dealer Adapt API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Car/{id}": {
      "get": {
        "tags": [
          "Car"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Car"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Car"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Car"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Car": {
      "get": {
        "tags": [
          "Car"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Car"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Car"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Car/{id}/cover-image": {
      "put": {
        "tags": [
          "Car"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CarFilter/{id}": {
      "get": {
        "tags": [
          "CarFilter"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/CarFilter/filter": {
      "post": {
        "tags": [
          "CarFilter"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CarFilterModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Car": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "make": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "model": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "mileage": {
          "format": "int32",
          "maximum": 2147483647,
          "minimum": 0,
          "type": "integer"
        },
        "price": {
          "format": "currency",
          "maximum": 1000000,
          "minimum": 0,
          "type": "number"
        },
        "color": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "vin": {
          "maxLength": 17,
          "minLength": 0,
          "type": "string"
        },
        "averageYearlyMaintenanceCost": {
          "format": "double",
          "minimum": 0,
          "type": "number"
        },
        "description": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "year": {
          "format": "int32",
          "maximum": 2500,
          "minimum": 1886,
          "type": "integer"
        },
        "isNew": {
          "type": "boolean"
        },
        "coverImageId": {
          "format": "uuid",
          "type": "string"
        },
        "pictures": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Picture"
          }
        },
        "listOfPictureUrls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "CarFilterModel": {
      "type": "object",
      "properties": {
        "makes": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "models": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "priceFrom": {
          "format": "double",
          "type": "number"
        },
        "priceTo": {
          "format": "double",
          "type": "number"
        },
        "colors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "mileageUpTo": {
          "format": "int32",
          "type": "integer"
        },
        "maintenanceCostFrom": {
          "format": "double",
          "type": "number"
        },
        "maintenanceCostTo": {
          "format": "double",
          "type": "number"
        },
        "yearFrom": {
          "format": "int32",
          "type": "integer"
        },
        "yearTo": {
          "format": "int32",
          "type": "integer"
        },
        "isNew": {
          "type": "boolean"
        },
        "keyword": {
          "type": "string"
        }
      }
    },
...
    "Picture": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "url": {
          "format": "uri",
          "type": "string"
        },
        "carId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
  }
}